Further abstraction
Moved Play button definition to a constant file for easier modification and cross-scene use. This required a version of Vector2 that didn't need to be instanced, so VectorPos was created. Unlike Phaser's Vector2, this only contains x and y with no additional functionality.

Renamed Text and InteractiveText to match export
