+++ src/interfaces/UI/DefaultImage.ts
@@ -1,7 +1,9 @@
+import VectorPos from "../universal/VectorPos";
+
 // A normal image with no additional properties for interactivity
 export default interface DefaultImage {
     texture: string;
     frame?: string | number | undefined;
-    offset: Phaser.Math.Vector2;
-    scale: Phaser.Math.Vector2;
+    offset: VectorPos;
+    scale: VectorPos;
 }
