+++ src/interfaces/UI/InteractiveText.ts
@@ -1,13 +1,8 @@
-import Phaser from "phaser";
-
-// Interface for Text
-// It has separate styles
-// for each state: normal, hover, and clicked
-export default interface Text {
-    text: string;                                           // Contents of the text
-    offset: Phaser.Math.Vector2;                            // Offset of the text, applied on add
-    origin: Phaser.Math.Vector2;                            // The origin of the text (useful for centering/alignment)
-    normalStyle: Phaser.Types.GameObjects.Text.TextStyle;   // Style used when not being interacted with
-    hoverStyle: Phaser.Types.GameObjects.Text.TextStyle;    // Style used when pointer hovers over the text
-    clickedStyle: Phaser.Types.GameObjects.Text.TextStyle;  // Style used when pointer is held down over text
+// Interface with the prototypes for interactive text
+export default interface InteractiveText {
+    pointerDown(text: Phaser.GameObjects.Text): void;
+    pointerUp(text: Phaser.GameObjects.Text): void;
+    pointerMove(text: Phaser.GameObjects.Text): void;
+    pointerOver(text: Phaser.GameObjects.Text): void;
+    pointerOut(text: Phaser.GameObjects.Text): void;
 }
